move the core plugin to the root and update internals on core plugin loading
