+++ plugin.yaml
@@ -1,0 +1,27 @@
+name: "@lando/core"
+registry:
+  app:
+    app: "./components/app"
+    minapp: "./components/minapp"
+  engine:
+    # docker-colima: "./components/docker-colima"
+    docker-desktop: "./components/docker-desktop"
+    docker-engine: "./components/docker-engine"
+  plugin-installer:
+    docker-plugin-installer: "./components/docker-plugin-installer"
+  storage:
+    file-storage: "./components/file-storage"
+  legacy:
+    tasks:
+      config: "./tasks/config"
+      destroy: "./tasks/destroy"
+      info: "./tasks/info"
+      list: "./tasks/list"
+      logs: "./tasks/logs"
+      poweroff: "./tasks/poweroff"
+      rebuild: "./tasks/rebuild"
+      restart: "./tasks/restart"
+      share: "./tasks/share"
+      start: "./tasks/start"
+      stop: "./tasks/stop"
+      version: "./tasks/version"
