--- plugins/core/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@lando/core",
-  "description": ""
-}
