--- plugins/core/plugin.yaml
@@ -1,16 +1,0 @@
-name: "@lando/core"
-registry:
-  legacy:
-    tasks:
-      config: './tasks/config'
-      destroy: './tasks/destroy'
-      info: './tasks/info'
-      list: './tasks/list'
-      logs: './tasks/logs'
-      poweroff: './tasks/poweroff'
-      rebuild: './tasks/rebuild'
-      restart: './tasks/restart'
-      share: './tasks/share'
-      start: './tasks/start'
-      stop: './tasks/stop'
-      version: './tasks/version'
