+++ scripts/add-cert.sh
