+++ scripts/load-keys.sh
