+++ scripts/log.sh
