+++ scripts/user-perm-helpers.sh
