+++ scripts/user-perms.sh
