+++ tasks/destroy.js
