+++ tasks/stop.js
