+++ tasks/version.js
