+++ utils/find-plugins.js
@@ -12,11 +12,12 @@ module.exports = (dir, depth = 1) => {
   // if dir doesnt exist then return []
   if (!fs.existsSync(dir)) return [];
 
+  // if depth 0 then only look in teh dir passed in
+  const dirs = depth === 0 ? [dir] : readdirSyncAbsDir(dir);
   // list of files that indicate we have a plugin
   const pConfigFiles = ['plugin.js', 'plugin.yaml', 'plugin.yml', 'package.json'];
   // rescurse through dirs until we are good
-  return readdirSyncAbsDir(dir)
-  .map(dir => {
+  return dirs.map(dir => {
     // return if path contains a plugin
     if (pConfigFiles.some(file => fs.existsSync(path.join(dir, file)))) return dir;
     // otherwise recurse if depth allows
